+++ examples/06-dev/demo/project/src/test-6/index/index.js
@@ -2,8 +2,6 @@
  * preserve comments
  */
 
-
-
 require('bootstrap/dist/css/bootstrap.css');
 require('./css/index.less');
 require('./css/index.css');
@@ -27,5 +25,5 @@ require('bootstrap');
 const $ = require('jquery');
 
 $.get('/src/test/index/data/index/?key1=value1&key2=value2', {}, res => {
-    console.log("get['/src/test/index/data/index']", res);
+  console.log("get['/src/test/index/data/index']", res);
 });