+++ examples/06-dev/prettier.js
@@ -1,0 +1,4 @@
+
+require('../../util/change_cwd')(__dirname + '/demo');
+
+require('../../util/exec')('lila prettier test-6/index');