+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.1 [2018-06-21]
 
+- Add `prettierOptions` option.
+- Add `prettier` command.
 - Add `styleLintOptions, styleLintFix` option.
 - Add `stylelint` command.
 - Add `esLintOptions, esLintFix` option.